pyproject.toml
[tool.poetry]
name = "bar-coin"
version = "0.1.0"
description = "Mobile-base Automated Coin Counter and Segregation for Coin Businesses Management"
authors = ["BAR-COIN Team"]
readme = "README.md"
packages = [{include = "bar_coin"}]

[tool.poetry.dependencies]
python = "^3.11"
streamlit = "^1.29.0"
pyserial = "^3.5"
pandas = "^2.1.0"
plotly = "^5.17.0"
bcrypt = "^4.1.0"
python-dotenv = "^1.0.0"
streamlit-authenticator = "^0.2.3"
streamlit-option-menu = "^0.3.6"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.0"
pytest-cov = "^4.1.0"
black = "^23.9.0"
flake8 = "^6.1.0"
mypy = "^1.6.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 88
target-version = ['py311']

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true 